+++ EssentialFeed/EssentialFeedTests/Feed Cache/XCTestCase+FeedStoreSpecs.swift
@@ -1,0 +1,69 @@
+//
+//  XCTestCase+FeedStoreSpecs.swift
+//  EssentialFeed
+//
+//  Created by Cristian Felipe Patiño Rojas on 7/4/25.
+//
+import XCTest
+import EssentialFeed
+
+extension FeedStoreSpecs where Self: XCTestCase {
+    @discardableResult
+    func insert(_ cache: (feed: [LocalFeedImage], timestamp: Date), to sut: FeedStore) -> Error? {
+        let exp = expectation(description: "Wait for cache insertion")
+        var insertionError: Error?
+        sut.insert(cache.feed, timestamp: cache.timestamp) { error in
+            insertionError = error
+            exp.fulfill()
+        }
+        
+        wait(for: [exp], timeout: 1.0)
+        return insertionError
+    }
+    
+    @discardableResult
+    func deleteCache(from sut: FeedStore) -> Error? {
+        let exp = expectation(description: "Wait for cache deletion")
+        var receivedError: Error?
+        sut.deleteCachedFeed {
+            exp.fulfill()
+            receivedError = $0
+        }
+        
+        wait(for: [exp], timeout: 1.0)
+        return receivedError
+    }
+    
+    func expect(
+        _ sut: FeedStore,
+        toRetrieve expectedResult: RetrieveCachedFeedResult,
+        file: StaticString = #file,
+        line: UInt = #line
+    ) {
+        let exp = expectation(description: "Wait for cache retrieval")
+        
+        sut.retrieve { retrievedResult in
+            exp.fulfill()
+            switch (expectedResult, retrievedResult) {
+            case (.empty, .empty), (.failure, .failure):
+                break
+            case let (.found(expectedFeed, expectedTimestmap), .found(retrievedFeed, retrievedTimestamp)):
+                XCTAssertEqual(expectedFeed, retrievedFeed, file: file, line: line)
+                XCTAssertEqual(expectedTimestmap, retrievedTimestamp, file: file, line: line)
+            default:
+                XCTFail("Expected to retrieve \(expectedResult), but got \(retrievedResult)", file: file, line: line)
+            }
+        }
+        wait(for: [exp], timeout: 1.0)
+    }
+    
+    func expect(
+        _ sut: FeedStore,
+        toRetrieveTwice expectedResult: RetrieveCachedFeedResult,
+        file: StaticString = #file,
+        line: UInt = #line
+    ) {
+        expect(sut, toRetrieve: expectedResult, file: file, line: line)
+        expect(sut, toRetrieve: expectedResult, file: file, line: line)
+    }
+}
